Fix parcel structure

+++ parcel.go
@@ -1,13 +1,13 @@
 package paack
 
 type Parcel struct {
-	Barcode          string `json:"barcode"`
-	Height           int    `json:"height"`
-	Length           int    `json:"length"`
-	Type             string `json:"type"`
-	VolumetricWeight int    `json:"volumetric_weight"`
-	Weight           int    `json:"weight"`
-	Width            int    `json:"width"`
+	Barcode          string  `json:"barcode"`
+	Height           int     `json:"height"`
+	Length           int     `json:"length"`
+	Type             string  `json:"type"`
+	VolumetricWeight int     `json:"volumetric_weight"`
+	Weight           float32 `json:"weight"`
+	Width            int     `json:"width"`
 }
 
 type Parcels struct {
